Move express server to status module
