+++ bot.ts
@@ -7,11 +7,8 @@ import {
     EmbedBuilder
 } from 'discord.js';
 import { CronJob } from 'cron';
-import express from 'express';
-import bodyParser from 'body-parser';
 
 // Modules
-import { BuildStatusUpdateReq, formatCommitShort, formatPiStatus, statusToColor } from './modules/status';
 import { fetchAndUpdateScoreboard, scoreboard } from './modules/scoreboard';
 import { Command, CommandGroup, getAllCommands } from './util/commands';
 
@@ -19,12 +16,7 @@ import { Command, CommandGroup, getAllCommands } from './util/commands';
 import {
     ATTACK_FORUM_CHANNEL_ID,
     ATTACK_NOTIFY_CHANNEL_ID,
-    DESIGN_REPO_URL,
-    EXPRESS_PORT,
-    FAILURE_CHANNEL_ID,
-    SCOREBOARD_NOTIFY_CHANNEL_ID,
-    STATUS_CHANNEL_ID,
-    STATUS_MESSAGE_ID
+    SCOREBOARD_NOTIFY_CHANNEL_ID
 } from './config';
 
 
@@ -147,75 +139,6 @@ export async function updateInfoForTeam(name: string, ip: string, portLow: numbe
 
     await attackThread.send({ embeds: [resEmbed] });
 }
-
-async function updateBuildStatus(req: BuildStatusUpdateReq) {
-    const channel = client.channels.cache.get(STATUS_CHANNEL_ID);
-    if (!channel?.isSendable())
-        return console.error('[BUILD] Could not find build status channel!');
-
-    const message = channel.messages.cache.get(STATUS_MESSAGE_ID)
-        || await channel.messages.fetch(STATUS_MESSAGE_ID)
-        || channel.lastMessage;
-
-    const queueStatus = req.build.queue.map((d, i) => `${i + 1}. ${formatCommitShort(d)}`).join('\n')
-        || '*No commits queued.*'
-    const piStatus = req.test.activeTests.map((s, i) => `${i + 1}. ${formatPiStatus(s)}`).join('\n');
-    const buildStatus = req.build.active
-        ? formatCommitShort(req.build.active)
-        : '*No commits loaded.*'
-
-    const color = req.status
-        ? statusToColor(req.status)
-        : '#27272a'
-
-    const statusEmbed = new EmbedBuilder()
-        .setTitle('Secure design build status')
-        .setDescription(`**Status:** ${req.status || 'N/A'}`)
-        .addFields(
-            { name: 'Pis', value: piStatus },
-            { name: 'Building:', value: buildStatus },
-            { name: 'Queued:', value: queueStatus }
-        )
-        .setColor(color)
-        .setTimestamp()
-
-    // Report build failures to the appropriate channel
-    if (
-        (req.update.type === 'BUILD' || req.update.type === 'TEST')
-        && req.update.state.result === 'FAILED'
-    ) {
-        const runHref = `${DESIGN_REPO_URL}/actions/runs/${req.update.state.commit.runId}`;
-
-        const failureChannel = client.channels.cache.get(FAILURE_CHANNEL_ID);
-        const failureEmbed = new EmbedBuilder()
-            .setTitle(`${req.update.type === 'BUILD' ? 'Build' : 'Tests'} failed for commit`)
-            .setColor(0xb50300)
-            .setDescription(`[\`${req.update.state.commit.hash.slice(0, 7)}\`]: ${req.update.state.commit.name} (@${req.update.state.commit.author})\n[[Jump to failed workflow]](${runHref})`)
-            .setTimestamp()
-
-        if (failureChannel?.isSendable())
-            failureChannel.send({ embeds: [failureEmbed] })
-    }
-
-    if (!message?.editable) return channel.send({ embeds: [statusEmbed] });
-    return message.edit({ embeds: [statusEmbed] });
-}
-
-const server = express();
-
-server.use(bodyParser.json());
-server.post('/', async (req, res) => {
-    console.log(`[BUILD] Received webhook push:\n${JSON.stringify(req.body)}`);
-    try {
-        await updateBuildStatus(req.body);
-        res.status(200).json({ ok: true });
-    } catch {
-        res.status(400).json({ ok: false });
-    }
-});
-server.listen(EXPRESS_PORT, () => {
-    console.log(`[BUILD] Started express server on port ${EXPRESS_PORT}`);
-});
 
 client.once('ready', async () => {
     client.commands = new Collection();
